+++ Chanlun.Lib/Memory/ChanCalculateResultCache.cs
@@ -4,12 +4,12 @@ namespace Chanlun.Lib.Memory;
 
 public static class ChanCalculateResultCache
 {
-    public static void Add(float key, ChanCalculateResult value)
+    public static void Add(decimal key, ChanCalculateResult value)
     {
         ChanMemory.Add(key,nameof(ChanCalculateResultCache), value);
     }
 
-    public static ChanCalculateResult? Get(float key)
+    public static ChanCalculateResult? Get(decimal key)
     {
         return  ChanMemory.Get<ChanCalculateResult>(key,nameof(ChanCalculateResultCache));
     }
